refactor(model): Use integer instead of float for the frequencies in CPU topology

+++ src/smartwatts/model/cpu_topology.py
@@ -37,7 +37,7 @@ class CPUTopology:
     This class stores the necessary information about the CPU topology.
     """
 
-    def __init__(self, tdp: int, freq_bclk: float, ratio_min: int, ratio_base: int, ratio_max: int):
+    def __init__(self, tdp: int, freq_bclk: int, ratio_min: int, ratio_base: int, ratio_max: int):
         """
         Create a new CPU topology object.
         :param tdp: TDP of the CPU in Watt
@@ -52,28 +52,28 @@ class CPUTopology:
         self.ratio_base = ratio_base
         self.ratio_max = ratio_max
 
-    def get_min_frequency(self) -> float:
+    def get_min_frequency(self) -> int:
         """
         Compute and return the CPU max efficiency frequency.
         :return: The CPU max efficiency frequency in MHz
         """
         return self.freq_bclk * self.ratio_min
 
-    def get_base_frequency(self) -> float:
+    def get_base_frequency(self) -> int:
         """
         Compute and return the CPU base frequency.
         :return: The CPU base frequency in MHz
         """
         return self.freq_bclk * self.ratio_base
 
-    def get_max_frequency(self) -> float:
+    def get_max_frequency(self) -> int:
         """
         Compute and return the CPU maximum frequency. (Turbo-Boost included)
         :return: The CPU maximum frequency in MHz
         """
         return self.freq_bclk * self.ratio_max
 
-    def get_supported_frequencies(self) -> List[float]:
+    def get_supported_frequencies(self) -> List[int]:
         """
         Compute the supported frequencies for this CPU.
         :return: A list of supported frequencies in MHz
